ci(lint): add linter to check we run all tests
- Add a script to check that a `suite_test.go` exists file whenever there a `test.go` file
- Fix files to make them pass the lint

Fix #1492
