+++ pkg/core/xds/inspect/suite_test.go
@@ -1,0 +1,11 @@
+package inspect_test
+
+import (
+	"testing"
+
+	"github.com/kumahq/kuma/pkg/test"
+)
+
+func Test(t *testing.T) {
+	test.RunSpecs(t, "XDS inspect Suite")
+}
