+++ pkg/xds/envoy/tls/tls_suite_test.go
@@ -1,0 +1,11 @@
+package tls_test
+
+import (
+	"testing"
+
+	"github.com/kumahq/kuma/pkg/test"
+)
+
+func Test(t *testing.T) {
+	test.RunSpecs(t, "Envoy TLS Suite")
+}
