+++ pkg/xds/generator/secrets/generator_suite_test.go
@@ -1,0 +1,11 @@
+package secrets_test
+
+import (
+	"testing"
+
+	"github.com/kumahq/kuma/pkg/test"
+)
+
+func Test(t *testing.T) {
+	test.RunSpecs(t, "Secrets Generator Suite")
+}
