+++ pkg/xds/generator/secrets/testdata/secrets/envoy-config-dataplane.golden.yaml
