+++ pkg/xds/generator/secrets/testdata/secrets/envoy-config-egress.golden.yaml
