+++ pkg/xds/generator/secrets/testdata/secrets/envoy-config-zipkin.golden.yaml
