+++ pkg/xds/template/proxy_template_suite_test.go
@@ -1,0 +1,11 @@
+package template_test
+
+import (
+	"testing"
+
+	"github.com/kumahq/kuma/pkg/test"
+)
+
+func Test(t *testing.T) {
+	test.RunSpecs(t, "ProxyTemplate Suite")
+}
